requirements.txt
asgiref==3.5.0
Django~=4.1.3
django-filter==2.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Pillow==9.1.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2

djangorestframework-jwt~=1.11.0